+++ bricks/_common/qstrdefs.h
@@ -9,10 +9,6 @@ Q(stop button pressed)
 Q(Li-ion)
 #endif
 
-#if PYBRICKS_PY_BLE
-Q(pybricks.ble)
-#endif
-
 #if PYBRICKS_PY_EV3DEVICES
 Q(pybricks.ev3devices)
 #endif
