+++ bricks/_common/sources.mk
@@ -22,7 +22,6 @@ CONTIKI_SRC_C = $(addprefix lib/contiki-core/,\
 # Pybricks modules
 
 PYBRICKS_PYBRICKS_SRC_C = $(addprefix pybricks/,\
-	ble/pb_module_ble.c \
 	ble/pb_type_broadcast.c \
 	common/pb_type_battery.c \
 	common/pb_type_charger.c \
