--- pybricks/ble.h
@@ -1,17 +1,0 @@
-// SPDX-License-Identifier: MIT
-// Copyright (c) 2018-2020 The Pybricks Authors
-
-#ifndef PYBRICKS_INCLUDED_PYBRICKS_BLE_H
-#define PYBRICKS_INCLUDED_PYBRICKS_BLE_H
-
-#include "py/mpconfig.h"
-
-#if PYBRICKS_PY_BLE
-
-#include "py/obj.h"
-
-extern const mp_obj_type_t pb_type_Broadcast;
-
-#endif // PYBRICKS_PY_BLE
-
-#endif // PYBRICKS_INCLUDED_PYBRICKS_BLE_H
