+++ tests/pup/ble/ble1_prime.py
@@ -9,7 +9,7 @@ Description: Tests broadcast receive and transmit.
 from pybricks.hubs import PrimeHub
 from pybricks.parameters import Color
 from pybricks.tools import wait
-from pybricks.ble import Broadcast
+from pybricks.experimental import Broadcast
 
 
 hub = PrimeHub()
