+++ tests/pup/ble/ble2_send.py
@@ -9,7 +9,7 @@ Description: Tests broadcast receive and transmit.
 
 from pybricks.parameters import Color
 from pybricks.tools import wait
-from pybricks.ble import Broadcast
+from pybricks.experimental import Broadcast
 
 from umath import pi
